fraud_detection_system/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
SQLAlchemy==2.0.23
alembic==1.12.1
pyodbc==5.0.1
fdb==2.0.2
firebird-driver==1.10.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Validation & Serialization
pydantic==2.5.0
pydantic-settings==2.1.0
email-validator==2.1.0

# WebSocket & Real-time
python-socketio==5.10.0
websockets==12.0

# Caching & Message Queue
redis==5.0.1
celery==5.3.4
kombu==5.3.4
flower==2.0.1

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7
sentry-sdk[fastapi]==1.38.0

# HTTP Client
httpx==0.25.2
aiofiles==23.2.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
faker==20.1.0

# Development Tools
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.5.0

# Circuit Breaker & Resilience
py-breaker==0.7.2
tenacity==8.2.3

# Data Processing
pandas==2.1.3
numpy==1.26.2
python-dateutil==2.8.2

# API Documentation
fastapi-limiter==0.1.5

# Excel/PDF Export
openpyxl==3.1.2
reportlab==4.0.7
xlsxwriter==3.1.9